pyproject.toml
[project]
name = "langchain-prompt-protection"
version = "1.0.0"
description = """\
An example CLI tool in Python that demonstrates how to integrate Pangea \
services into a LangChain app to capture and filter what users are sending to \
LLMs.\
"""
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "click>=8.1.7",
    "langchain-openai>=0.2.0",
    "langchain>=0.3.0",
    "pangea-sdk>=4.2.0",
    "pydantic>=2.9.2",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.11.2",
    "ruff>=0.6.7",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # ruff defaults
    "E4",
    "E7",
    "E9",
    "F",

    "I",   # isort
    "PTH", # flake8-use-pathlib
]
